software/bios: move mdio to libliteeth.
