+++ litex/soc/software/bios/commands/cmd_mdio.c
@@ -5,7 +5,7 @@
 
 #include <generated/csr.h>
 
-#include <base/mdio.h>
+#include "mdio.h"
 
 #include "../command.h"
 #include "../helpers.h"
