+++ litex/soc/software/libliteeth/mdio.c
