+++ litex/soc/software/libliteeth/mdio.h
